requirements.txt
argparse
ipaddress
colorama
pycryptodomex